dbt/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
dbt-core==1.9.2
dbt-snowflake==1.9.1
snowflake-connector-python==3.13.2
python-dotenv==1.0.1

